requirements.txt
bs4 
colorama
py-solc 
